trace-cmd: Fix python plugins not resolving symbols
When writing a python plugin, it would fail to load with the following
error:

failed loading /root/.trace-cmd/plugins/my_plugin.py
Traceback (most recent call last):
  File "<string>", line 1, in <module>
  File "/usr/local/lib/trace-cmd/python/tracecmd.py", line 22, in <module>
    from ctracecmd import *
ImportError: /usr/local/lib/trace-cmd/python/ctracecmd.so: undefined symbol: tep_unload_plugins

This was due to a change to the placement of the tracing files, and the
libtracecmd.a was separated from the LIBS makefile variable. Not only that,
a new library was added libtraceevent.a which also was required by
ctracecmd.so. By adding these libraries to the build of ctraceccmd.so, the
python plugins should work again.

Fixes: e480376f62d0 ("trace-cmd: Move libtraceevent *.c files in lib/traceevent")
Reported-by: Johannes Berg <johannes@sipsolutions.net>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>