kernel-shark-qt: Fix a bug in the plugin actions execution
Plugin-provided actions are executed when loading the data. These
actions can be used to modify the content of the kshark_entries
generated by a given event type and we consider the case of having
multiple plugin-provided actions per event type. However, the code
that handles the case of multiple actions per-event has a bug. The "if"
was introduced with the idea that only the last per-event action will
modify the KS_PLUGIN_UNTOUCHED flag of the entry, but it misbehaves in
the case of a single per-event action in the list, followed by actions
for other event types.

Fixes: c70227ebb264 ("kernel-shark-qt: Add Plugin event handlers to session.")
Signed-off-by: Yordan Karadzhov <ykaradzhov@vmware.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>