kernel-shark-qt: Fix formatting of fixed-point numbers to not skip zeros
When a timestamp that represents 1.00001 is converted to (for example)
seconds and microseconds this results in seconds=1 and useconds=10. To
format that into a string requires the format %lu.%06lu to yield
"1.000010". Some code parts used %lu.%lu however which results in "1.10"
which is wrong.

This was noticed because I failed to find a significant location in a trace
provided by a colleague as the reported timestamp was wrong :-|

This change fixes all occurrences of this problem found by
$(grep -E '\.%[^0s]').

Reviewed-by: Yordan Karadzhov <ykaradzhov@vmware.com>
Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>