kernel-shark-qt: Handle the case when the range of the model is too small
The size of the bin must be equal or greater than 1, hence the range
must be equal or greater than the number of bins. When this condition
is not satisfied, the range has to be extended, instead of just leaving
the model unset.

This fixes an issue where a trace.dat of one event will cause a segfault.

Link: http://lore.kernel.org/linux-trace-devel/20190115101857.25857-1-ykaradzhov@vmware.com

Reported-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
Fixes: f97e31f00 ("kernel-shark-qt: Introduce the visualization model used by the Qt-based KS")
Signed-off-by: Yordan Karadzhov <ykaradzhov@vmware.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>