kernel-shark: Remove testing of "success" field of wakeup events
Since 2011 (or Linux version 3.0) the sched wakeup event "success" field has
been hardcoded to true (or 1). There's no reason to waste cycles checking it
in any newer kernel, which is most of them. Just ignore it. Even when
running on older kernels which may have a success=0 value, the algorithm
should still work.

Link: http://lore.kernel.org/linux-trace-devel/20190211012153.848572065@goodmis.org

Reviewed-by: Yordan Karadzhov <ykaradzhov@vmware.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>