kernel-shark: Do not copy the Upper Overflow bin when shifting forward
ksmodel_shift_forward() copies the mapping indexes of all overlapping
bins of the model starting from bin "0" of the new histo up to bin
"histo->n_bins - n". Then the mapping index of the old Upper Overflow
bin is considered to be the mapping index of first non-overlapping bin,
which is wrong. It is wrong because in ksmodel_set_upper_edge() the
value of "histo->max" is considered inside the range of the model hence
the Upper Overflow bin starts at "histo->max + 1" but the first
non-overlapping bin will start at exactly "histo->max".

Link: http://lore.kernel.org/linux-trace-devel/20190213161216.14438-3-ykaradzhov@vmware.com

Fixes: f97e31f00 ("kernel-shark-qt: Introduce the visualization model used by the Qt-based KS")
Signed-off-by: Yordan Karadzhov <ykaradzhov@vmware.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>