trace-cmd: Have tracecmd_msg_data_send() return zero on zero size
If tracecmd_msg_data_send() is passed a size of zero, the loop is not
entered, nothing is sent, but the ret variable is not initialized and
garbage can be sent out. The option code can call this with a size of zero,
which can cause an error report.

Just don't do anything in this case, and return a success.

Link: http://lore.kernel.org/linux-trace-devel/20190321194127.4ac47d99@gandalf.local.home

Reviewed-by: Slavomir Kaslev <kaslevs@vmware.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>