trace-cmd: Add trace-cmd report --version to extract version info
Now that the trace.dat file can save the version of the executable that was
used to create it, save that information. This will be useful if a user has
a problem with a trace.dat file. It can let the developers know what version
was used to create it. We need a way to extract this information. Adding a
new "--version" to trace-cmd report will display the saved version in the
trace.dat file if it exists.

Link: http://lore.kernel.org/linux-trace-devel/20190610212706.2505bffe@oasis.local.home

Reviewed-by: Slavomir Kaslev <kaslevs@vmware.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>