trace-cmd: Install all headers needed by kernelshark to proper paths
The trace-cmd build system installs headers for building kernelshark,
but it is missing trace-seq.h and trace-filter-hash.h.

It additionally installs event-parse.h to the toplevel "trace-cmd"
include directory, but the generated header includes it as
"traceevent/event-parse.h" so the kernelshark build fails.

This patch installs all the headers needed to build kernelshark to the
appropiate directories.

Link: http://lore.kernel.org/linux-trace-devel/20190720030344.1991393-4-chutzpah@gentoo.org

Signed-off-by: Patrick McLean <patrick.mclean@sony.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>