trace-cmd: Make a global tracecmd_quick_hash() instead of a local knuth_hash()
As the 32 bit Knuth algorith produces the same results as the small
sizes[1], create a single algorithm that takes in a @bits parameter that
will return a masked result. And replace the local version of knuth_hash()
used by the trace-cmd filter code. This will also be used to remove other
copies of knuth_hash().

[1] https://lore.kernel.org/r/20190829114913.5df4ced9@gandalf.local.home

Link: http://lore.kernel.org/linux-trace-devel/20190920152024.567157833@goodmis.org

Reviewed-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>