kernel-shark: Increase the size of the task hash
When loading a data file that contained 100,000s of tasks, using a 256
bucket size hash crippled it. By increasing the hash to 2^16 (65536) it
solves the issue (still small enough not to waste too much memory).

Also switched to the tracecmd_quick_hash() which is basically the same
as the local knuth_hash() function in libkshark.c.

Link: http://lore.kernel.org/linux-trace-devel/20190828140016.3ce1be4f@gandalf.local.home
Link: http://lore.kernel.org/linux-trace-devel/20190920152024.729716704@goodmis.org

Reviewed-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>