trace-cmd: Add initial infrastructure for trace-cmd specific plugins.
Create a new directory lib/trace-cmd/plugins, where the libtracecmd
plugins will be placed. Plugins compilation is integrated in
the trace-cmd build process. New plugins are installed in
$(HOME)/.local/lib/trace-cmd/plugins or $(libdir)/trace-cmd/plugins directories.
Current python_plugin.so is part of libtraceevent plugins, so the python
plugins path is changed from
$(HOME)/.trace-cmd/python to $(HOME)/.local/lib/traceevent/python
and
$(libdir)/trace-cmd/python to $(libdir)/traceevent/python

Added an initial set of APIs for working with trace-cmd plugins.

Link: http://lore.kernel.org/linux-trace-devel/20191007114947.17104-4-tz.stoyanov@gmail.com

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
[ Removed redundant declarations of tracecmd_disable_sys_plugins and tracecmd_disable_plugins ]
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>