trace-cmd: Check if ftrace file exists, before writing in it.
Depending of the kernel configuration, some ftrace files are optional
and may not exist. This should not confuse trace-cmd, as it is a valid
case. One such case is tracing_max_latency file, when CONFIG_TRACER_MAX_TRACE
or CONFIG_HWLAT_TRACER are not set.
A check is added in write_instance_file() to ensure the file exist before trying
to write in it.

Link: http://lore.kernel.org/linux-trace-devel/20191018145759.15335-2-tz.stoyanov@gmail.com

Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=205241
Reported-by: Valentin Schneider <valentin.schneider@arm.com>
Tested-by: Valentin Schneider <valentin.schneider@arm.com>
Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>