trace-cmd: Fix latency format to show latency
When trace-cmd report -l is used, the latency format is missing. This is
due to the show data routine not using TEP_PRINT_LATENCY. Need to change
the logic slightly where different parameters is sent to tep_print_event()
when latency_format is set, instead of just passing a different format
file.

Link: http://lore.kernel.org/linux-trace-devel/20191122165537.26f89069@oasis.local.home

Fixes: 150d479b623a ("libtraceevent, perf tools: Changes in tep_print_event_* APIs")
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>