tools lib traceevent: Add builtin handler for trace_marker_raw
When something is written into trace_marker_raw, it goes in as a binary. But
the printk_fmt() of the event that is created (raw_data)'s format file only
prints the first byte of data:

  print fmt: "id:%04x %08x", REC->id, (int)REC->buf[0]

This is no very useful if we want to see the full data output.

Implement the processing of the raw_data event like it is in the kernel.

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>