trace-cmd: Have /etc paths honor $(prefix) if not the default
As some files need to be placed in the root directory to take effect (like
the bash_completion) and not the normal /usr/local directory, they ignore
the $(prefix) and install directly. Unfortunately, this means that if
someone does a build like "make O=/tmp/mydir install" those files will still
try to be installed in the root "/etc" directory, and that will not work if
the user installing does not have permission.

Have the "/etc" directory change if prefix is something other than
"/usr/local".

Reported-by: Julia Lawall <julia.lawall@inria.fr>
Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=206277
Fixes: 77bdcb9b2 ("trace-cmd: Install trace-cmd.bash when installing")
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>