trace-cmd: Fix "multiple definition" compiler errors
When compiling with gcc and option "-fno-common", there are
errors because of multiple global variables with the same names.
The "-fno-common" gcc option is turned on by default on gcc 10, so
trace-cmd build fails on this compiler.

Proposed solution: set the variables to be static, as their scope is
only the current C file.

Link: http://lore.kernel.org/linux-trace-devel/20200210085010.7974-1-tz.stoyanov@gmail.com

Reported-by: Ziqian SUN (Zamir) <sztsian@gmail.com>
Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=206467
Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>