trace-cmd: Have instances include a copy of its output file
Instead of having a global "output_file", store the output file in the
instance itself. That way it can be used instead of having to pass around
a global variable.

This also fixes a bug when an output file option is used, and it can break
the recording of the guest, as the tsync currently looks for the "default"
recording name. Instead, it can now look at the output file stored in the
instance (which will hold the guest name).

Link: http://lore.kernel.org/linux-trace-devel/20200309190526.886340672@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>