trace-cmd: Fix check for trigger file in create_event()
When a trigger is configured on an event with trace-cmd "-R"
option there is a check in create_event() if the trigger file
exist, using the stat() call. It returns 0 if the file exists
or -1 in case of an error. The check of the stat()'s return code
is for positive number, which is always false and errors are never
detected.

Link: http://lore.kernel.org/linux-trace-devel/20200325140829.184651-2-tz.stoyanov@gmail.com

Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=206737
Fixes: 30ee2ad72 trace-cmd: Add trigger option '-R' to trace-cmd record/start
Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>