trace-cmd: Fix double free error when triggers are configured on multiple events
When using trace-cmd "-e sched:* -R ..." command to set triggers to the all events
of given subsystem, trace-cmd fails with a double free error.
The problem is in create_event() function, where the config is expanded to the all
events. The same "event->trigger" pointer is copied to all events. Later, when the
trigger is applied, the update_event() frees this pointer multiple times.

Link: http://lore.kernel.org/linux-trace-devel/20200325140829.184651-3-tz.stoyanov@gmail.com

Fixes: 30ee2ad72 ("trace-cmd: Add trigger option '-R' to trace-cmd record/start")
Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>