trace-cmd: Add new APIs to pair and unpair tracing peers
When tracing host and one or more guest machines at the same time,
guest and host are tracing peers. There is information in both trace
files, related to host PID to guest vCPU mapping, timestamp synchronization
and other. This information is useful when opening files at the same time and
merging the events. When the host is set as a tracing peer to the guest, then
the timestamps of guest's events are recalculated to match the host event's time
Two new APIs are intrduced to control pairing of tracing peers:

int tracecmd_pair_peer(struct tracecmd_input *handle,
		       struct tracecmd_input *peer);
int tracecmd_unpair_peer(struct tracecmd_input *handle);

Link: http://lore.kernel.org/linux-trace-devel/20200413082334.164158-2-tz.stoyanov@gmail.com

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>