trace-cmd: Add new tracefs APIs for checking if a ftrace file / directory exists
Some ftrace files and directories are optional, depending on specific
kernel configuration or version. It is a good practice to check if
the file / directory exist, before trying to access it.
There are a lot of places in trace-cmd implementation with such checks.
The new libtracefs APIs can be used for this, they are ftarce instance aware:
 bool tracefs_file_exist(struct tracefs_instance *instance, char *name);
 bool tracefs_dir_exist(struct tracefs_instance *instance, char *name);

Link: http://lore.kernel.org/linux-trace-devel/20200414155604.31401-2-tz.stoyanov@gmail.com

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>