trace-cmd: Do not enable timestamps synchronization by default
When a timestamps synchronization data is detected in the trace file options,
and if this data is valid - timestamps correction is enabled by default. This
behavior was for PoC stage of the patches and should not be used in the final
version.
Timestamps synchronization should be enabled only if the file is opened
together with a file from the same tracing session. A new APIs were introduced
and already merged upstream for pairing peers from the same session. There is a
KernelShark patch which uses those new APIs to pair peers from the same session,
when loading the tracing files.
The default behavior, indented to be used for the PoC should be removed.

Link: http://lore.kernel.org/linux-trace-devel/20200424092749.638101-1-tz.stoyanov@gmail.com

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>