trace-cmd: Explicitly state what trace-cmd report -f does
The current man page and usage of trace-cmd report is confusing when it
comes to the -f option. This option will list all the functions in the
trace.dat file that have a address to name mapping. As it currently is
stated, it may confuse users in thinking that it will list the functions
that have been traced. But this is not the case.

Link: http://lore.kernel.org/linux-trace-devel/20200708203017.175213296@goodmis.org

Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=207069
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>