trace-cmd: Fix "trace-cmd extract" temp files path
During the introduction of instance's output_file copy:

  3a206ca ("trace-cmd: Have instances include a copy of its output file")

The extract path has been omitted, causing the temp files created to be
written in the same directory using the null "output_file" of the
instance, to create "(null).cpuX" files. If this is executed in a
directory that is not writable (like /sys/kernel/tracing) or does not
have enough space to hold the temp files, then it will fail to write.

Link: https://lore.kernel.org/linux-trace-devel/1601542769-443324-1-git-send-email-vincent.donnefort@arm.com

Signed-off-by: Vincent Donnefort <vincent.donnefort@arm.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>