trace-cmd: Update .gitignore
The .gitignore file is somewhat out of date. Update the main file as well as
adding a .gitignore in the tracecmd and kernel-shark directories.

Remove the exec names that use to be included in the top level directory,
and place them into the .gitignore in the directories that may build them.

Ignore files caused by running "patch".

Add *.data for perf files that may be in the directory.

Add .pc and patches/ to ignore quilt directories.

Add #*# to ignore emacs meta files.

Add a .gitignore file in kernel-shark to include all the files that are
built by the kernelshark build.

Link: https://lore.kernel.org/linux-trace-devel/20201210022943.377786365@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>