trace-cmd: Have cunit check in Makefile write to /dev/null
The cunit test in the Makefile does a compile test to check if the cunit
package is installed or not. But it does not have a "-o" option, and thus
will write the result to "a.out" leaving the "a.out" in the source
directory. By adding "-o /dev/null" to the compile check, it will write the
output to /dev/null and not leave an extra executable around.

Link: https://lore.kernel.org/linux-trace-devel/20201210125858.0de1f116@gandalf.local.home

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>