trace-cmd: Do not install libtracefs nor libtraceevent libraries
Now that both libtraceevent and libtracefs have their own repos and are
packaged separately from trace-cmd, do not install the libraries even if
they do not exist on the system. Anything that depends on those libraries
are require to get them from their corresponding repos and not depend on
trace-cmd for them from now on.

Link: https://lore.kernel.org/linux-trace-devel/20201211171014.706323506@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>