trace-cmd: Do not build shared libraries for libtracefs and libtraceevent
As libtraceevent and libtracefs are now packaged separately and are built
from their own corresponding repos, the local copy that trace-cmd uses is
only in the case that the libraries are not present on the local system.
In this case, trace-cmd should build its wont static library and link to
that. This will be deprecated in the future and trace-cmd will eventually
depend on the external libraries for libtracefs and libtraceevent.

Link: https://lore.kernel.org/linux-trace-devel/20201211171014.867885861@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>